New main page
Created a new main page so that you can select either the IMS side or
POS side of the application.
